\name{gQTLBase-package}
\alias{gQTLBase-package}
\alias{gQTLBase}
\alias{matprint}
\docType{package}
\title{
\packageTitle{gQTLBase}
}
\description{
\packageDescription{gQTLBase}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gQTLBase}
\packageIndices{gQTLBase}

Purpose is to define infrastructure on a comprehensive archive
of eQTL, mQTL, dsQTL, etc., association statistics.

Package will complement gQTLStats.  geuvStore is a basic
illustration relative to GEUVADIS paper.

\code{\link[ff]{matprint}} is exported from package ff.
}
\author{
\packageAuthor{gQTLBase}

Maintainer: \packageMaintainer{gQTLBase}
}
%\references{
%}
%\examples{
%}
